Optimized API Mapper
