+++ PentestMapper.py
@@ -18,6 +18,13 @@ from urlparse import urlparse
 import time, csv, sys, os, base64, zlib
 from threading import Thread, Event, Lock
 from java.lang import Integer, String
+
+
+from pentestmapper.autolog import Autologclas  
+from pentestmapper.APIMapper import LogEntry, Table, CustomTableModel
+from pentestmapper.autosave import Autosaveclas 
+
+
 csv.field_size_limit(sys.maxsize)
 
 
@@ -137,8 +144,18 @@ class BurpExtender(IBurpExtender, ITab, IContextMenuFactory, AbstractTableModel,
         self.comboBox1.addItem("Pending")
         self.comboBox1.addItem("In Progress")
         self.comboBox1.addItem("Completed")
+
+
+        # creating a message editor from burp to show request 
+        self.requestViewer = callbacks.createMessageEditor(None, True)
+        self.responseViewer = callbacks.createMessageEditor(None, True)
+
         # creating a table with custom model for api mapper
-        self.logTable = Table(self)
+
+
+        self.apimappertable = CustomTableModel(self)
+
+        self.logTable = Table(self, self.apimappertable)
         
 
         # allowed colum size for api mapper tab/table
@@ -163,7 +180,7 @@ class BurpExtender(IBurpExtender, ITab, IContextMenuFactory, AbstractTableModel,
         self.scrollPane2 = JScrollPane(self.logTable)
         
         self.scrollPane2.getViewport().setView((self.logTable))
-        self.sorter2 = TableRowSorter(self);
+        self.sorter2 = TableRowSorter(self.apimappertable);
         self.logTable.setRowSorter(self.sorter2)
 
        
@@ -187,9 +204,7 @@ class BurpExtender(IBurpExtender, ITab, IContextMenuFactory, AbstractTableModel,
    
        
 
-        # creating a message editor from burp to show request 
-        self.requestViewer = callbacks.createMessageEditor(None, True)
-        self.responseViewer = callbacks.createMessageEditor(None, True)
+        
         
        
        #adding the tapped pane to create request and test cases 
@@ -961,72 +976,6 @@ class BurpExtender(IBurpExtender, ITab, IContextMenuFactory, AbstractTableModel,
             if searchedvalue in items:
                 self.comboBox.addItem(items)
     
-    # part of custom table model to get total number of row in the table, it will check the data in the list amd will return when called     
-    def getRowCount(self):
-        try:
-            return len(self._log)
-        except:
-            return 0
-
-    # this is required to work with AbstractTableModel, this will show return the total columns in API Mapper
-    def getColumnCount(self):
-        return 6
-
-    # this is required to work with AbstractTableModel, this will show the columns Name for API Mapper table
-    def getColumnName(self, columnIndex):
-        if columnIndex == 0:
-            return "SR"
-        if columnIndex == 1:
-            return "URL"
-        if columnIndex == 2:
-            return "Method"
-        if columnIndex == 3:
-            return "Post Body"
-        if columnIndex == 4:
-            return "Functionality Name"
-        if columnIndex == 5:
-            return "Status"
-        return ""
-
-    # this is required to work with AbstractTableModel, this will show the data in the columns from the list for API Mapper table
-    def getValueAt(self, rowIndex, columnIndex):
-        self.totalrow = self.logTable.getRowCount()
-        if rowIndex < self.getRowCount() and columnIndex < self.getColumnCount():
-            logEntry = self._log[rowIndex]
-            if columnIndex == 0:
-                return str(rowIndex + 1)
-            if columnIndex == 1:
-                return logEntry._url
-            if columnIndex == 2:
-                return logEntry._method
-            if columnIndex == 3:
-                return logEntry._postbody
-            if columnIndex == 4:
-                return logEntry._FunctionalityName
-            if columnIndex == 5:
-                return logEntry._status
-            return ""
-        else:
-            self.callbacks.printError("Table is empty")
-
-    ''' 
-    part of custom table for api mapper, it will get the edited value from function column and will save it in the class based list 
-    for the selected row
-
-    '''
-    def setValueAt(self, value, rowIndex, columnIndex):
-        
-        if rowIndex < self.getRowCount() and columnIndex == 4:
-                logEntry = self._log[rowIndex]
-                logEntry._FunctionalityName = value
-
-        if rowIndex < self.getRowCount() and columnIndex == 5:
-                logEntry = self._log[rowIndex]
-                logEntry._status = value
-
-
-        else:
-            self.callbacks.printError("Table is empty")
    
     #function that will show a option to send request to extension in proxy/repeater etc
     def createMenuItems(self, invocation):
@@ -1071,7 +1020,7 @@ class BurpExtender(IBurpExtender, ITab, IContextMenuFactory, AbstractTableModel,
             row = len(self._log)
             column = self.logTable.getColumnCount()
             self._log.append(LogEntry(self.sr2, self.url, self.method, self.body, self.requestinst, '', '', self.responseinst, ''))
-            self.fireTableRowsInserted(row, row)
+            self.logTable.getModel().fireTableRowsInserted(row, row)
             self._lock.release()
 
 
@@ -1290,157 +1239,10 @@ class CustomDefaultTableModelHosts2(DefaultTableModel):
         else:
             return 1
 
-# Extedning the JTable for API Mapper TaB
-class Table(JTable):
-    def __init__(self, extender):
-        self._extender = extender
-        self.setModel(extender)
-        self.setRowSelectionAllowed(True)
-        self.setAutoCreateRowSorter(True)
-        
-    def getColumnClass(self, column):
-        columnClasses = [Integer, String, String, String, String,String]
-        return columnClasses[column]
-            
-    # Only last colum is allowed for editing    
-    def isCellEditable(self, row, column):
-        if column == 4 or 5:
-            return 1
-        else:
-            return 0
-    
-    # function that will change the request and test case for selected row
-    def changeSelection(self, row, col, toggle, extend):
-        self.modelRowIndex = self.convertRowIndexToModel(row) 
-        self._extender._currentlySelectedLogTableRow = self.modelRowIndex
-        logEntry = self._extender._log[int(self.modelRowIndex)]
-        self._extender.requestViewer.setMessage(
-            logEntry._requestResponse, True)
-        self._extender.responseViewer.setMessage(
-            logEntry._response, True)
-        text = logEntry._TestCases
-        self._extender.testcases.setText(logEntry._TestCases)
-        self._extender._currentlyDisplayedItem = logEntry._requestResponse
-        JTable.changeSelection(self, row, col, toggle, extend)
 
 
 
-##Class to handle auto save
-class Autosaveclas(Thread):
-    def __init__(self,handlingoutput):
-        self._handlingoutput = handlingoutput
-       
-        Thread.__init__(self)
-        self._stop = Event()
-
-    def run(self):
-        
-        #self._handlingoutput = handlingoutput
-        while not self.stopped():
-            if self._handlingoutput.Autosavechecker == True:
-        
-                if os.path.isdir(str(self._handlingoutput.path)):
-
-                    if self._handlingoutput.logTable.getRowCount() > 0:
-                        #self.path = fileLoad.getPath()
-                        fname = "APIMapper"+"."+"csv"
-                        fnameWithPath = os.path.join(self._handlingoutput.path,fname)
-                        if os.path.exists(fnameWithPath):
-                            os.remove(fnameWithPath)
-                        self._handlingoutput.callbacks.printOutput("Saving the API Mapper output")
-                        with open(fnameWithPath, 'wb') as loggerdata:
-                            writer = csv.writer(loggerdata)
-                            for logEntry in self._handlingoutput._log:
-                                writer.writerow([str(logEntry._sr), str(logEntry._url) ,str(logEntry._method) , base64.b64encode(zlib.compress(logEntry._postbody.encode('utf-8'))) ,str(logEntry._FunctionalityName) , base64.b64encode(zlib.compress(logEntry._requestResponse.encode('utf-8'))) ,str(logEntry._TestCases), base64.b64encode(zlib.compress(logEntry._response.encode('utf-8'))) ,str(logEntry._status)])
-                                 
-                            loggerdata.close()
-                    else:
-                        self._handlingoutput.callbacks.printOutput("Skipping the API Mapper, Table is empty")
-                    if self._handlingoutput.dataModel2.getRowCount() > 0:
-                        fname2 = "Vulnerability"+"."+"csv"
-                        fnameWithPath2 = os.path.join(self._handlingoutput.path,fname2)
-                        if os.path.exists(fnameWithPath2):
-                            os.remove(fnameWithPath2)
-                        self._handlingoutput.callbacks.printOutput("Saving the Vulnerability output")
-                        totalrow = self._handlingoutput.dataModel2.getRowCount()   
-                        with open(fnameWithPath2, 'wb') as vulnerabilitydata:
-                            writer = csv.writer(vulnerabilitydata)
-                            for row in range (0, totalrow):
-                                url = self._handlingoutput.dataModel2.getValueAt(row,0)
-                                paramter = self._handlingoutput.dataModel2.getValueAt(int(row),1)
-                                Vulnerability = self._handlingoutput.dataModel2.getValueAt(int(row),2)
-                                Severity = self._handlingoutput.dataModel2.getValueAt(int(row),3)
-                                #self.callbacks.printOutput(str(url))
-                            
-                            #self.callbacks.printOutput(str(logEntry._sr))
-                                writer.writerow([str(url), str(paramter) ,str(Vulnerability),str(Severity)])
-                            vulnerabilitydata.close()
-                    else:
-                        self._handlingoutput.callbacks.printOutput("Skipping the  Vulnerability, Table is empty")
-                else:
-                    self._handlingoutput.autosavepath.setText("Output Directory doesn't exist")
-                time.sleep(int(self._handlingoutput.time) * 60)
-
-    def stop(self):
-        self._stop.set()
-
-    def stopped(self):
-        return self._stop.isSet()
 
 
 
-#logger entry <- Class based python list to store the data for API Mapper tab/tabble
-class LogEntry:
-    def __init__(self, sr, url, method, postbody, requestResponse, FunctionalityName, TestCases, response, status):
-        self._sr = sr
-        self._url = url
-        self._method = method
-        self._postbody = postbody
-        self._requestResponse = requestResponse
-        self._FunctionalityName = FunctionalityName
-        self._TestCases = TestCases
-        self._response = response
-        self._status = status
 
-# Class to handle auto log requests from proxy
-class Autologclas(IProxyListener):
-    def __init__(self,handlingoutput):
-        self._handlingoutput = handlingoutput
-        #self._handlingoutput.callbacks.registerHttpListener(self)
-
-
-    def processProxyMessage(self,messageIsRequest, message):
-
-        if self._handlingoutput.Autologcheck == True:
-            if not messageIsRequest:
-                if self._handlingoutput.callbacks.isInScope(self._handlingoutput.helpers.analyzeRequest(message.getMessageInfo()).getUrl()):
-
-                    req = self._handlingoutput.helpers.analyzeRequest(message.getMessageInfo())
-                    self.url = self._handlingoutput.helpers.analyzeRequest(message.getMessageInfo()).getUrl()
-                    response = message.getMessageInfo().getResponse()
-                    responseInfo = self._handlingoutput.helpers.analyzeResponse(response)
-
-                    # Find out if image
-                    self.responsetype = responseInfo.getInferredMimeType()
-                    self.responsetype2 = responseInfo.getStatedMimeType()
-                    self.exludelist = self._handlingoutput.Excludefiletextfield.getText().split(",")
-                    
-                    if (self.responsetype.upper() not in self.exludelist) and (self.responsetype2.upper() not in self.exludelist):
-                    #if self.responsetype.upper() not in self.exludelist:
-                        
-                        self.method = req.getMethod()
-                        self.requestinst = self._handlingoutput.helpers.bytesToString(message.getMessageInfo().getRequest())
-                        getody = req.getBodyOffset()
-                        self.body = self.requestinst[getody:len(self.requestinst)]
-                        self.responseinst = self._handlingoutput.helpers.bytesToString(message.getMessageInfo().getResponse())   
-                        rowss = self._handlingoutput.logTable.getRowCount()
-                        self.sr2 = str((rowss + 1))
-                        
-                        self._handlingoutput._lock.acquire()
-                        row = len(self._handlingoutput._log)
-                        self._handlingoutput._log.append(LogEntry(self.sr2, self.url, self.method, self.body, self.requestinst, '', '', self.responseinst,''))
-                        self._handlingoutput.fireTableRowsInserted(row, row)
-                        self._handlingoutput._lock.release()
-            
-        else:
-            self._handlingoutput.callbacks.removeProxyListener(self)
